+++ sample-project/replace_android_tokens.sh
@@ -1,3 +1,0 @@
-sed -i '' 's/\$APPBOY_ANDROID_API_KEY/7ef72382-b92f-4258-9be9-9e19e92f3bf1/g' platforms/android/res/values/appboy.xml
-sed -i '' 's/\$APPBOY_ANDROID_PUSH_REGISTRATION_ENABLED/true/g' platforms/android/res/values/appboy.xml
-sed -i '' 's/\$APPBOY_ANDROID_GCM_SENDER_ID/901477453852/g' platforms/android/res/values/appboy.xml